changed some integer variables to long long int as indices can get beyond the int range

+++ get_nodeData.c
@@ -38,9 +38,9 @@ int get_nodeData(char * filename, fpos_t startPos, int nNodes, struct node ** no
 		(*nodeData)[n].expansionFactor = atof(pch);
 		for(i=1;i<27;i++) {
 			pch = strtok(NULL," ");
-			if(i==1) (*nodeData)[n].nodeIndex = atoi(pch);
-			if(i==3) (*nodeData)[n].descendentIndex = atoi(pch);
-			if(i==6) (*nodeData)[n].hostIndex = atoi(pch);
+			if(i==1) (*nodeData)[n].nodeIndex = atoll(pch);
+			if(i==3) (*nodeData)[n].descendentIndex = atoll(pch);
+			if(i==6) (*nodeData)[n].hostIndex = atoll(pch);
 			if(i==9) (*nodeData)[n].nodeMass = atof(pch);
 			if(i==12)(*nodeData)[n].scaleRadius = atof(pch);
 			if(i==17)(*nodeData)[n].position[0] = atof(pch);
